requirements.txt
google-cloud-aiplatform
vertexai
tenacity
ffmpeg-python
psycopg2-binary

